Big fix for broken sections, table form updates, etc.
Fixed an issue where sections aren't being added in the right place; without
some notion of object identity for sections, it's not generally decidable where
to insert a section (since we can't compare the contents of a section, as that
might change, we really only know that the number of sections is different). The
solution is to add unique identifiers to each section, by which sections will be
quotiented.

RAFTableSection now has mutable and immutable interfaces. This obviates the need
for RAFTableSectionMoment, which had grown to have an identical interface to
RAFTableSection.

Furthermore, RAFTableSection has a signal of its moments, which is useful for
getting an immutable view of changing mutable data.