Fixed broken behavior for raf_cast and signal replaying
-rac_cast for forms needs to spit out the data, not the form itself.
It's kind of a gross cludge that we define it globally for NSObject as
the identity, and then give it different semantics for forms, so ideas
on replacements would be appreciated. But at least as designed, this
change corrects the obviously broken previous behavior.

Furthermore, the totalDataSignal and validationSignal should not replay
the whole thing, but only the last value. This allows us to use
totalDataSignal.first, etc.